stm:snd:fli7610 Power down AATV before configuring
This patch is from Bugzilla 18095. It's purpose is to power down the AATV
before we configure it to ensure that the AATV is in a known state.

This also fixes an issue of no sound from the head phones socket on the HDK
board when the board has been warm booted.

Signed-off-by: Sudhesh Nayak <sudhesh.nayak@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>