net: fix the nwhwconfig invocation
After the latest merge (commit: 72eab04683f8d) the network has some
problems due to the MAC address passed to the stmmac device driver.  In
fact it is random generated each boot and this creates problems due to
routing table update process, for example.

The NWHWCONFIG has been erroneously moved in the drivers/net/Makefile to
top of the file.  It should stay at the end otherwise the
nwhwconf_init() fails because no dev interface has been probed by the
net stack when nwhwconfig_init() is called. However, it has to be called
before ip_auto_config() which will actually open the device. So it needs
to be called after all ethernet devices have been probed but before the
device is opened. Unfortunately there is no init level that allows this.

Instead of hacking the Makefile again and moving the nwhconfig to the
end, this patch changes the init level for the nwhwconfig from
device_initcall() to late_initcall(). This still relies on link ordering
to function correctly, as it now relies on the link order in the top
level makefile to ensure this is called before ip_auto_config().  This
works because "drivers" comes before "net" in the Makefile.

Hopefully this should help avoid this kind of problem in the future when
merging from the main-stream, as the top level makefile is unlikely to
change this order.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: David McKay <david.mckay@st.com>