arm: smp: stm: define scu_base_addr as a SoC specific variable
Currently scu_base_addr is defined as a simple function just returning
the virtual base address of the ARM Snoop Control Unit.
The implementation returns the hard-coded value of the SCU base address
of stih415 (orly) for all STM ARM platforms.
It is not an ideal situation as newer chips could have the SCU mapped
at different physical address.

This patch uses a global variable 'scu_base_addr' that must be defined and
initialised in the soc specific port implementation with the proper value.
(arch/arm/mach-stm/soc-<soc_name>.c)

A similar approach is used for example in the 'highbank' chip by Calxeda.

Signed-off-by: Nunzio Raciti <nunzio.raciti@st.com>
Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: David McKay <david.mckay@st.com>