stm: sata: Update the pm_runtime support
This patch updates the pm_runtime support to be really
used on-demand.

In order to ensure that the PM_runtime doesn't suspend the sata
controller there is a call to pm_runtime_get() in the probe function. In
this manner the sata_host is enabled and the pm_runtime subsystem
maintains the sata_host as enabled.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>