stm: clk: remove ALWAYS_ENABLED flag from the generic code
This patch changes how the ALWAYS_ENABLED flag is used.
From now on the clk framework will use only the usage_counter field
to enable/disable on the fly every clock.
The ALWAYS_ENABLED flags is moved to an stm specific file,
and permanently enabled clocks are flagged during registration.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>