stm:i2c: Allow fastmode to be set via plat data
Add a flag to the stm I2C platform data which can be used to initialise
the device's fastmode flag. This allows boards which have an I2C bus
populated entirely with fastmode capable slave devices to set the flag
as part of their initialisation of the I2C device, and so remove the
need for later configuration using the sysfs device.

This models the way the PXA I2C driver sets its fast_mode flag.

By default all boards have their channels set to standard mode rather
than fastmode. It really intended to be mainly used on customer boards
which don't have the tuners on NIM slots as ST does on the majority of
it's boards.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Signed-off-by: David McKay <david.mckay@st.com>