stm:snd Fix uniperipheral config for layout1 transmission
This fix is for Bugzilla 16689.

For PCM, the REPEAT_CHL_STS bit of UNIPERIF_CONFIG register shall be set
to 0 (enabled) for PCM layout1. This means the same channel status is
transmitted on each pair of channels.

For HBRA, the REPEAT_CHL_STS bit of UNIPERIF_CONFIG register would be set
to 1 (disabled), although HBRA is not yet implemented.

Signed-off-by: Aziz Kehal <aziz.kehal@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>