stm:ir_stm: Send timeout event for last space.
This patch sends an timeout event to lirc codec to indicate an end of
signal. Without this patch any pulse miss in between a signal train will
leave the next signals undecoded.

Originally detected with STM-PVR-1 remote, while pressing keys if we move
remote away from the receiver some of the signals for a key press will
not be received by the driver and the irw application,
As irw fails to get a valid code for the pulse-spaces it continues to
take more pulses-spaces from next key press resulting in no key decode
till a timeout occurs.

With this patch a timeout event will re-sync the irw application and
let the follow on key press detected.

Reported-by: Wei HE <he.wei-amk@st.com>
Tested-by: Wei HE <he.wei-amk@st.com>
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>