stm:b2000: Improve resiliance to clock chip failure
The B2000 board has a OnSemi NB3N5573 clock generator chip (U31) to
provide the needed high accuracy PCI Express clock signal. For some as
yet unknown reason there seems to be a high failure rate on these chips.

This clock is needed for PCIe to function, however it is also needed to
talk to the Miphy which is also used for SATA. If this clock is not
present you will get a data abort from the ARM when attempting to talk
to the Miphy.

The only real fix for this is to replace the faulty chip. At the same
time you may need to change R131 to 1-2 and R130 to 2-3 depending on
your board revision to generate the correct 100Mhz clock.

What this patch does is only flip to the external clock if you have
CONFIG_PCI on. This means that if you have a board with a broken chip,
if you disable CONFIG_PCI you should be able to at least have working
SATA, though PCIe will never work unless the board is modified.

Signed-off-by: David McKay <david.mckay@st.com>