stm-ir: Fake first long-space for first key detection.
lircd expects an long space before any valid space-pulse train. As lirc
driver moved from 2.6 to 3.x one of the major changes was the fifos
are managed by the rc infrastructure in 3.x. 2.6 kernel lirc driver used
to cache the signals and send long space once the timeout is detected
and then send the cached signals.
This is not possible in 3.x kernel as stm-ir driver does not cache the
signals. stm-rc driver only sends a long space when a timeout is
detected, this happens only at the end of the signal train. So, If
stm-ir driver has to send a long space at start of every keypress, It
has to fake the first long space. The subsequent long-spaces will be
form the previous key-press last timeout space.

Without patch :
       key press 1                            	key press 2
       ___  ___  ___                         ___  ___  ___
       | |  | |  | |                         | |  | |  | |
       | |  | |  | |   long space 0          | |  | |  | |  long space 1
  _____| |__| |__| |_________________________| |__| |__| |_____________

After the patch :
                Key press1                            	 key press 2
                   ___  ___  ___                      ___  ___  ___
                   | |  | |  | |                      | |  | |  | |
   fake-long-space | |  | |  | |   long space   0     | |  | |  | | LS 1
___________________| |__| |__| |______________________| |__| |__| |_____

Without this patch first key press is lost after box boot's up.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>