arm: Expose PHYS_OFFSET as an ELF note.
This patch fulfills a request to expose the value of PHYS_OFFSET -
effectively the physical address the kernel is linked for - in the ELF
file for use by bootloaders etc.

This is only meaningful if CONFIG_ARM_PATCH_PHYS_VIRT isn't set, in
which case the note is given the value 0xffffffff.

See Bugzilla #18514.

Signed-off-by: Chris Smith <chris.smith@st.com>