stm:snd Fix channel status bits and validity bits update
This is related to Bugzilla 19642.

The problem is that the channel status bits and validity bits are not
being updated correctly. The cause is using a bitwise AND instead of a
logical AND in a conditional statement when determining if the bits have
changed and if the uniperipheral is running.

Resolution is to just update the bits when they change regardless of
whether the uniperipheral is running or not.

Signed-off-by: Sanjay Jha <sanjay.jha@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>