stm_sata: Handle spurious irqs better
For some unknown reason it appears that we can get interrupts from the
SATA block even when they are explicitly masked out. In particular we
seem  to see a CMD_GOOD interrupt being raised even when this is masked
out. This seems to be hardware problem during the early probe stages.

In older version of this IP, the actual ATA interrupt status was not
available, so the code basically had to assume that if it the interrupt
was raised it was an ATA interrupt. The problem is that when we call
into the ATA code from this spurious irq there is no actual ATA
transaction outstanding.  The code assumes there has to be a transaction
in progress if we call into the ATA code.

Later versions of the IP have the ATA irq information available, so this
patch uses this information. It should be available for all recent ST
socs. So now we only call into the ATA code when we know it actually is
a real ATA interrupt.

Signed-off-by: David McKay <david.mckay@st.com>