stm: Rework stm_soc identification
This replaces the previous ad-hoc SoC identification functions. The
main change is the introduction of a new stm/soc.h header file, and
the replacement of the stm_soc_type() function with a number of
stm_soc_is_xxx() macros. This follows the pattern of other SoCs
supported in the ARM tree, and allows source code which tests which
SoC they are running on to be optimised by the compiler when this
is known.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>