stm:dma STMicroelectronics FDMA dmaengine API audio extensions
This patch adds audio extensions to the dmaengine API. The audio extensions
consists of new API to allow the addition of a 'parking mode' for the
uniperipheral player.

The audio extensions allow nearly all of the 'parking mode' implementation
to exist outside of the uniperipheral player, seriously reducing it's
impact on that driver.

The audio extensions add the following API:
- dma_audio_parking_config()
  This allows configuration of the 'parking mode' parameters.
- dma_audio_parking_enable()
  This switches from running a transfer to the 'parking mode'. To exit the
  'parking mode' merely start a new transfer or terminate all transfers.
- dma_audio_is_parking_active()
  Returns if 'parking mode' is active
- dma_audio_prep_tx_cyclic()
  Specially prepare a dmaengine transfer taking into account that the
  transfer may be asked to enter 'parking mode' as quickly as possible.

The audio extensions have been tested with a modified version of the
uniperipheral player driver that will be submitted later.

Signed-off-by: John Boddie <john.boddie@st.com>