stm:dma Update to FDMA firmware loading
There is a problem with the FDMA firmware loading. If a channel initiates
a firmware load and it fails, then the firmware state is still marked as
'loading'. If another channel then checks for the firmware being loaded,
it sees the firmware state is 'loading' and waits for it to complete. Only
the firmware is not loading and we end up waiting forever for a
non-existant load to complete.

This patch updates the FDMA firmware loader to include an error state so
that if the firmware fails to load, any subsequent check returns error.

Signed-off-by: John Boddie <john.boddie@st.com>