stm:snd Use clock aliases in uniperipheral player driver
This patch modifies the uniperipheral player driver to use clock aliases
instead of a clock specified in the platform device configuration. This
makes the uniperipheral player driver consistent with the PCM and SPDIF
player drivers.

As part of this work the name of the uniperipheral drivers have changed
from snd_uniperif_player to snd_uni_player and from snd_uniperif_reader to
snd_uni_reader. The reason being that the clock aliasing only supports a
device with a name under 20 characters.

The SPDIF RX converter is also updated to use clock aliases.

Signed-off-by: John Boddie <john.boddie@st.com>