stm:snd:pm Initial PM implementation for uniperipheral drivers
This patch relates to Bugzilla 18004 and provides an initial implementation
of power management for the uniperipheral player and reader drivers based
on what is done for other ALSA drivers.

If parking mode is active when the uniperipheral player is suspended, then
the parking mode is terminated. Parking mode is not re-enabled on resume
as parking mode can only be switched to when stopping running audio.

Signed-off-by: John Boddie <john.boddie@st.com>