stm: fli7610 and fli76xxhdk: Fix SSC problems
This fixes a couple of problems with the fli7610 SSC support:
 - the SCL and SDA pins for I2C3 were reversed
 - the second SPI port is on SBC SSC1 not SBC SSC0.

In addition the I2C3 bus was not configured on the fli76xxhdk board.

The FLI7610 documentation uses a slightly odd numbering scheme, where
the ports are numbered from 1 not 0, and distinguish between whether the
port is intended to be used for I2C or SPI. So also introduce a full
set of macros enumerating the different ports which correspond to the
names in the documentation so that the user doesn't need to know
the SSC numbers at all.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Acked-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>