stm:fdma Fix for FDMA driver error interrupt path BUG_ON
This patch relates to Bugzilla 20903.

There is a race condition in the FDMA driver error interrupt handling. On
receiving an error interrupt the FDMA channel will be paused and the
current active descriptor processing is passed off to a error tasklet.

Before the error tasklet can run, an interrupt will fire to indicate the
FDMA channel is paused. This will also process the current active
descriptor as a completion, meaning that when the error tasklet finally
runs there is no active descriptor and the BUG_ON triggers.

This patch moves the FDMA channel pause calls into the error tasklet,
ensuring that the error tasklet completes before the next interrupt.
Similarly all of the completion code has been moved into the completion
tasklet. As an added benefit the FDMA driver no longer needs to take the
FDMA channel lock in interrupt context.

Signed-off-by: John Boddie <john.boddie@st.com>