stm:snd Do not activate parking mode on XRUN
This patch relates to Bugzilla 20903.

If parking mode is enabled, the uniperipheral player driver attempts to
activate parking mode everytime it is stopped, including when stopped from
an XRUN condition.

Switching to the parking buffer from an XRUN condition will usually make
the uniperipheral player driver unstable.

This patch sets a flag on an XRUN condition to ensure that when the
uniperipheral player stops it does not activate parking mode. When the
uniperipheral player starts, this flag is reset.

Signed-off-by: John Boddie <john.boddie@st.com>