stm: snd: Added the snd_stm_clk wrapper
This patch adds the snd_clk_clk API which is a wrapper
on top of standard clk API.

The snd_stm_clk wrapper tracks if a clock
was already enabled or not (showing to the clk framework
a balanced clk_enable/clk_disable); this is required
because the snd system is not guaranting
balanced clk_enable/clk_disable

Moreover with the snd_stm_clk wrapper there is no reason
to register 'virtual' clk in the clk_framework.

In the same patch all the occurrences of 'struct clk' (and the
related function) are replaced with snd_stm_clk.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>