stm: pm: add multiple table support
Recently several customers are asking to move only
one lmi into self-refresh.

This was impossible with the previous implementation because
the poke_table was hard-coded to support dual lmi.

This patch reworks the pm code to be able to use multiple tables
with different actions.
When the system is going in standby or in hom the final real
table is built in eram.

At the moment this patch only puts the framework in place, there
is no mechanism to use this functionality.

This commit was merged with 399e4f65b6e4e4e0e2c6d1392531c4f49e0f5824
by David McKay. That patch added LMI retention pin support.

Tested on b2000-stih415: both mem-standby and hom.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: David McKay <david.mckay@st.com>