stm: b2020: Add missing customization in uncompress.h for uImage generation
The uImage generated for b2020 is not bootable by u-boot because a
customization modification for this board is missing.

Before the kernel starts, some asm code is executed to uncompress the kernel
at the correct address. This code is calling the function get_uart_base()
from arch/arm/mach-stm/include/mach/uncompress.h to get ASC base address in
order to print the trace "Uncompressing Linux kernel...".
But this function needs to be modified to return the proper asc base address
for the uart device, which is board specific, otherwise it returns NULL and
the A9 keeps being reset.

See STLinux bug 21039

Signed-off-by: André Draszik <andre.draszik@st.com>