stm:fdma Explicit channel error state
This patch adds a new explicit channel state (STM_FDMA_STATE_ERROR) that is
set when the channel experiences an error. Whilst this is similar to the
STM_FDMA_STATE_STOPPING state, but there are a couple of instances where
the error state is handled differently.

The error tasklet has been modified to graciously handle the case where it
is called and the active descriptor list is empty.

The channel error state is now also returned when getting the tx_status.

Signed-off-by: John Boddie <john.boddie@st.com>