stm:snd: an i2c driver always needs to provide an id_table for probing to work
From Documentation/i2c/writing-clients:

The probe function is called when an entry in the id_table name field
matches the device's name. It is passed the entry that was matched so
the driver knows which one in the table matched.

Without the .id_table, .probe will never be called (the driver's
driver.name field is not taken into account for this).

See STLinux bug 19376

Signed-off-by: André Draszik <andre.draszik@st.com>