stm:mali: implement a new MALI GPU hardware configuration.
This implements the standard ARM approach for creating and
configuring the Mali GPU hardware.
To do this, its removes our standard platform resources
registration and create the config.h configuration file
containg the Mali GPU, cache and MMU definitions.
This approach is valid for the current and next Mali kernel
driver version.

Signed-off-by: Angelo Castello <angelo.castello@st.com>