stm-asc: Simplify the way early console info is passed to driver
In the existing code, stm-asc driver recevies info about early console
device in 3 different variables, which is a not nice. It just needs info
about memory and irq which is platform data for console device.

With this patch only one console device platform device variable is
passed to the stm-asc as part of early console setup.

This simplification also eases device tree integration.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>