stm: Update Newman/Orly configs to VMSPLIT_2G
This changes the defconfigs for all the Orly and Newman boards
to have the user space/kernel boundary at 2G/2G rather than 3G/1G.
The reasons for doing this are that some boards now have 1G of physical
memory, and that cannot be supported in a 3/1 split with using HIGHMEM.
Another reason is that the SHARK project requires the 2/2 split to
provide enough kernel virtual address space. It doesn't make any sense
to use a different split in the defconfig.

Signed-off-by: David McKay <david.mckay@st.com>