bpa2: Allow BPA2 partitions to be contiguous with kernel memory
Fix an effective off-by-one error which prevents BPA2 partitions which
are outside kernel logical memory from being contiguous with it.

end_pfn and min_low_pfn, which represent the upper bounds of the
memory regions being described, are actually the first page outside
the region, as they are start+length, not start+length-1.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>