fli76xxhdk01: fix defconfig
The Fli76xxhdk01 default defconfig is broken; by default the
CONFIG_MACH_STM_B2020 option is enabled and the Kernel is built for stih415.
This problem has been introduced with the commit:
45cb0cfdc25d29efe3e4992f415f455768278615

So this patch fixes this problem and turns-off the PM stuff becasue
untested and unsupported on this SoC. In fact, the following error is
raised when compiled with PM enabled.

LD      arch/arm/mach-stm/built-in.o
armv7-linux-ld: cannot find arch/arm/mach-stm/pm/built-in.o: No such file or directory

Kernel runs fine on FLI7650HDK3 board.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>