stm: stx7108: enable MAC RX COE
On new chips, there is the HW capability register to detect
if the RX COE is supported but on old chips (i.e. STx7108)
we need to pass this information from the platform.

W/o this patch the stmmac assumed that the 7108 MAC wasn't
able to perform the csum in HW so it was calculated in SW
for all the incoming frames.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>