stm: h415: clk: remove ddrclk set_rate
This patch removes the ddrclk_set_rate() function  to avoid the ddrclk
being turned-off while the CPU is running on DDR

This happens during every PM (standby, mem, hom, hibernation) resume.
During the resume the clk-framework guarantees the clks in memory are
aligned to the clk in hw. To do that it forces all the settings and
(when the .set_rate is available) the clk-framework does a set_rate on
each clk.

In case of the ddrclk the set_rate has to:
 - turn-off the clk
 - set the rate
 - turn-on the clk

During these steps the CPU stops working because the DDR has no
clock.

This patch totally removes the ddrclk_set_rate() function because in
any case this operations can not be performed in C-code.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>