stm:sata:Remove un-necessary delays after COMRESET and COMWAKE.
SATA speed negotiation procedure begins after the host receives the
COMWAKE from the device. But if the host sleeps for too long, it will
miss the (first) Fastest rate ALIGNS and result in negotiation at lower
speeds.

This patch removes the un-necessary delays in the reset procedure to
ensure the SATA speed is negotiated correctly.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>