net:stmmac: Add check if mdiobus is registered in stmmac_mdio_unregister()
This patch adds a basic check in stmmac_mdio_unregister() to see if mdio
bus registeration for this driver was actually sucessfull or not.

Use case here is, if BSP considers using mdio-gpio bus along with stmmac
driver by passing mdio_bus_data as NULL in platform data.  Call to
stmmac_mdio_register() with mdio_bus_data as NULL returns 0, which is a
considered sucessfull call form stmmac. Then again when we unload the
driver we just call stmmac_mdio_unregister(), this is were the actual
problem is. The stmmac-mdio code doesn't know if this instance of calling
that stmmac_mdio_register was actually successful.

So Adding a check in stmmac_mdio_unregister() is always safe.

Without this patch stmmac driver calls stmmac_mdio_register() from
stmmac_release() which segfaults as mii bus was never registered at the
first point.

Originally the this bug was found when unloading an stmmac driver
instance which uses mdio-gpio for smi access.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>