pcie_stm: Cope with for zero size io windows
In the recent merge to 3.4 (commit 72aeb6264) the IO address space was
changed to use the standard values of 0->64K. Look at the __io and
IO_SPACE_LIMIT macros in arch/arm/include/asm/io.h

Since ST's hardware doesn't support any form of io (inbi()/outb() etc),
the IO window size is set to be zero sized before it is passed into the
glue layer.

If we request a zero sized region from it will now fail, causing the
entire PCIe init routine to fail. This patch introduces an extra test to
cope with zero sized regions, if the io region is zero sized it will not
be requested. This now means that on our hardware the IO regions on the
card will be disabled which is a desirable property anyway.

Signed-off-by: David McKay <david.mckay@st.com>