stm: Change stxxxx_configure_ethernet parameter to use phy_interface_t
All the stxxxx_configure_ethernet() calls take a parameter which
describes the PHY interface mode. Previously this was a different enum
for each SoC, which then had to be converted into the standard
phy_interface_t. So switch to using the phy_interface_t directly.

The only mode which could not be represented this way was the
STx7108's distinction between GMII with and without a GTX
clock. However the mode without the GTX clock has never been used, and
was dropped from the STx7108 datasheet (ADCS 8183887) between
revisions A and B, the newer version of which states that a GTX clock
must always be used.

Note that all RGMII modes are specified as RGMII_ID, to automatically
configure the interface delays in the IC+1001 used on those boards.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>