net: Add basic reverse MII support
Some devices support a reverse MII mode, where the MAC signals are
inverted, making it appear as a PHY. This allows two MACs to be
connected back to back with no additional logic.

Add the basic support for this, a new entry in 'enum phy_interface_t'
and the corresponding change in modes accepted by of_get_phy_mode().

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>