stm: Resurrect SH4 support on STxH415
Some of the recent updates to the kernel have broken support for the
SH4 on the STxH415 and b2000, while the ARM was still working. These
changes get the SH4 working again.

Most of the changes simply involve reinstating code which was accidentally
deleted.

The only real change is to match the changes to the tmu driver, which
no longer has a parameter specifying the clock name, and now always
requests "tmu_fck". We used to specify this as "module_clk", but as
this has now gone, remove "module_clk" completely, as this was the
only remaining user. "CLKM_ICN_REG_10" is a guess as to the correct
clock to use, but appears to be the only 100MHz clock on the MPE41.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>