stm: pio-control: Always apply retiming
Previously we only applied retiming if the caller specified it. However
this means that for 'simple' pad configurations, like gpio, the previous
retiming is left unchanged, which may not be correct. So always apply
a retiming configuration if the pad support it, setting it into pass
mode if none is supplied. Conversely, if the user supplies a retiming
configuration and the pad doesn't support it, generate a warning.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>