stm: pio-control: Add the new dedicated pio retiming style
New chips are using a new retiming sysconf style where there is a
dedicated sysconf register for each GPIO's pin that has retiming.

This means that for each GPIO port at least 12 sysconf are
required and for this reason this patch is also increasing
the number of statically allocated sysconf.

Signed-off-by: Nunzio Raciti <nunzio.raciti@st.com>
Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>