stm: pio-control: Add a retime_style
To make it easier to add support for future SoCs which have different
configuration registers for the pad retiming, introduce a retime_style
field to the pio-control config. The first two retiming styles are
none (which allows us to remove the current no_retiming field) and
packed.

We also need to introduce a retime_pin_mask field, which can be
used when only some of the pads within a port implement retiming.
Currently this is always initialised to 0xff, for those ports which
support retiming, as for the packed style the unused fields are
always available and marked as reserved.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>