stm: clk: add alias for stmmaceth
This patch adds the clock_alias for the stmmaceth
renaming in some case stmmac_clk as stmmaceth that
is the driver name.
Recently the clk support has been reviewed in the stmmac
d.d. that fails if the clk_csr_i (System clock) is
not found. So we have to guarantee that this alias
is defined in all the ST platforms.
Indeed, in some case we should distinguish between
stmmaceth.0 and stmmaceth.1. In fact, on SoCs that
have two GMAC embedded we could use two different
clocks and the right alias should be stmmaceth.X.
This will be fixed in the future because it impacts
other platforms (not only ST).
Moreover, the clk_csr_i has been never used and
we have ever seen no issues because the range of the
clk_csr_i has been always fixed between 60 and 100 MHz
so no particular settings were ncessary to scale some
internal clocks (i.e. MDIO).

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>