rc/smt-ir: fix NULL pointer dereference in irq handler
This patch closes a small window of opportunity for the driver to crash
between irq_request and setting the ir device pointer.
In the ir_stm_probe() function the interrupt is enabled and further down in probe
we set the ir_dev->rdev = rdev.

So if an interrupt is raised in middle of these two events then there is
a possibility of this crash to happen.

This patch moves irq enable after rdev is set, so that if any sensitive
ir raises an interrupt it will have a valid rdev pointer.

Reference: https://bugzilla.stlinux.com/show_bug.cgi?id=21478
Signed-off-by: Jerome Brunet <jerome.brunet@st.com>
Tested-by-by: Jerome Brunet <jerome.brunet@st.com>
Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>