stm: stih415: fix RGMII when negotiated speed is 10
In case of RGMII and the speed is 10Mbit/s we set the
PHY clock to 2.5MHz and this is ok.
Unfortunately, while calling the clk_set_rate on
CLKS_ETH1_PHY (B2020) and we try to change the rate
to 2.5MHz this has no effect because the parent
CLKS_A0_PLL1 800.00MHz cannot provide this kind of
divided clocks.

So this patch routes, in case of the speed is reduced to
10, the PHY clock (for GMAC 0 and GMAC 1 in RGMII modes) to
another parent: in this specific case a good parent is
the 30MHz quartz.

For example, for the b2020, for the GMAC1 in RGMII mode when
speed is 10 the CLKS_ETH1_PHY will be routed from CLKS_A0_PLL1
to CLKS_A0_REF.

See Bugzilla Bug: 22226

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Cc: Francesco Virlinzi <francesco.virlinzi@st.com>