stm/arm/l2x0: Enable Instruction and Data prefetching.
This patch enables Instruction and Data prefetching in PL310.
Prefetch-related features can improve overall system performance.

There is no reason why these features should not be enabled, since the
revision of the PL310 used in all the ST chips is late enough such that
it is not afffected by the various prefetching errata such as 504326.

Errata 765569 is present in the hardware but the software is currently
not changing the PREFETCH_CTRL register so will not be affected. Just
don't set it to 23 :-)

However, the double linefill feature is still broken so please don't
enable that unconditionally (Errata 752271) in the future.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>
Signed-off-by: David McKay <david.mckay@st.com>