stm_i2c: Change fastmode sysfs file interface
Previously the speed of the i2c bus could be set to fastmode (400KHz) by
writing a "1" to the file called "fastmode" in the sysfs directory for
that particular i2c bus.

This patch changes this  to a file called "speed" which contains the
current bus speed in KHz.

This allows for running at non-standard speeds or higher standard speeds
such as the 1MHz FastMode+ in the future.

For example:

echo 400 > /sys/devices/platform/i2c-stm.0/i2c-0/speed

Will select 400KHz for bus zero.

In the current implementation, the only values accepted are 100 or 400.
Any other values written to the file will be ignored and the bus speed
will remain unchanged.

Signed-off-by: David McKay <david.mckay@st.com>