stm_i2c: Change platform data i2c bus speed selection to KHz
Previously fastmode (400KHZ) was selected by setting a field called
"fastmode" to 1 in the platform data. This patch changes this to be a
"speed" parameter, which is the desired speed of the bus in KHZ.

This increases future flexibility to allow non standard speeds and
faster standard speeds to be selected when supported in the future.

Signed-off-by: David McKay <david.mckay@st.com>