stm: stig125: add the ISVE support
This patch adds the ISVE platform fields for stig125 SoC.
At this stage, the queues #3, #4 and #5 are supported.
Although, only the former has been tested for communicating
A9 with eCM.
In the next future, it is likely the queue 7 will be added too.

For more information about the platform fields please see the
isve.txt Kernel document.

V2: moved from core to hispeed and added the new iface platform
field in order to provide custom names to the network interfaces.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>