stm:fdma Ignore IDLE to IDLE state changes
In theory, when stm_fdma_stop() is called around the same time at which a
transfer generates a completion event, the completion tasklet could handle
both the transfer completion and the state change associated with the stop
before the interrupt associated with the stop has actually scheduled the
completion tasklet.

This would mean the when the completion tasklet runs to handle the state
change associated with the stop, the state change has already happened.
This could result in a BUG_ON when moving from the IDLE state to the IDLE
state.

This patch ignores IDLE state to IDLE state transitions.

Signed-off-by: John Boddie <john.boddie@st.com>