stm:fdma Move IRQ processing from tasklets back into interrupt handler
This patch fixes both Bugzillas 21472 and 21563. It moves the processing
for error and complete interrupts out of individual tasklets into the main
interrupt handler. The active descriptor completion code from both the
error and complete tasklets has been combined into a single descriptor
complete function that now runs as a tasklet. Before freeing FDMA channel
resources, we first call tasklet_kill() on the descriptor complete tasklet.
This waits for the tasklet to finish exceuting, then removes it from the
pending queue.

For Bugzilla 21472, this stops IDLE to IDLE state transitions under heavy
system load where there is a delay in running the complete tasklet.
Previously if a transfer completed just before a stop was issued, rather
than the complete tasklet being scheduled to complete the transfer, then
scheduled again to process the stop IDLE state change, it was possible
both operations were being processed in the first tasklet schedule, so on
the second tasklet schedule the FDMA was already in the IDLE state.

For Bugzilla 21563, these changes prevent the FDMA channel resources from
being freed whilst the descriptor complete tasklet is running. Previously
if the uniperipheral closed whilst a complete was in progress, the FDMA
channel would be released then certain unipeirpheral state set to null. If
the complete tasklet then executed the transfer callback this uniperipheral
null state would be passed into ALSA causing a crash.

Signed-off-by: John Boddie <john.boddie@st.com>